+++ apps/ifala-backend/src/main/java/br/edu/ifpi/ifala/admin/Administrador.java
@@ -1,0 +1,51 @@
+package br.edu.ifpi.ifala.admin;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+
+/**
+ * Classe que representa um administrador no sistema.
+ *
+ * @author Renê Morais
+ */
+@Entity
+public class Administrador {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+  private String email;
+  private String senha;
+
+  /**
+   * Construtor padrão da classe Administrador.
+   */
+  public Administrador() {}
+
+  public Long getId() {
+    return id;
+  }
+
+  public void setId(Long id) {
+    this.id = id;
+  }
+
+  public String getEmail() {
+    return email;
+  }
+
+  public void setEmail(String email) {
+    this.email = email;
+  }
+
+  public String getSenha() {
+    return senha;
+  }
+
+  public void setSenha(String senha) {
+    this.senha = senha;
+  }
+
+}
