--- apps/ifala-backend/src/main/java/br/edu/ifpi/ifala/controllers/DenunciaController.java
@@ -1,5 +1,0 @@
-package br.edu.ifpi.ifala.controllers;
-
-public class DenunciaController {
-
-}
