+++ apps/ifala-backend/src/main/java/br/edu/ifpi/ifala/denuncia/DenunciaRepository.java
@@ -1,0 +1,11 @@
+package br.edu.ifpi.ifala.denuncia;
+
+/**
+ * Classe repositório para gerenciamento de entidades Denuncia. Responsável por operações de
+ * persistência e recuperação de denúncias no sistema.
+ *
+ * @author Renê Morais
+ */
+public class DenunciaRepository {
+
+}
