+++ apps/ifala-backend/src/main/java/br/edu/ifpi/ifala/denuncia/DenunciaService.java
@@ -1,0 +1,10 @@
+package br.edu.ifpi.ifala.denuncia;
+
+/**
+ * Classe de serviço responsável por manipular operações relacionadas a denúncias.
+ *
+ * @author Renê Morais
+ */
+public class DenunciaService {
+
+}
