--- apps/ifala-backend/src/main/java/br/edu/ifpi/ifala/entities/enums/Categorias.java
@@ -1,19 +1,0 @@
-package br.edu.ifpi.ifala.entities.enums;
-
-public enum Categorias {
-    CELULAR ("Celular"),
-    DROGAS ("Drogas"),
-    BULLYING ("Bullying"),
-    VIOLENCIA ("Violência"),
-    OUTROS ("Outros");
-
-    private String descricao;
-
-    Categorias(String descricao) {
-        this.descricao = descricao;
-    }
-
-    public String getDescricao() {
-        return descricao;
-    }
-}
