--- apps/ifala-backend/src/main/java/br/edu/ifpi/ifala/repositories/DenunciaRepository.java
@@ -1,5 +1,0 @@
-package br.edu.ifpi.ifala.repositories;
-
-public class DenunciaRepository {
-
-}
