--- apps/ifala-backend/src/main/java/br/edu/ifpi/ifala/services/DenunciaService.java
@@ -1,5 +1,0 @@
-package br.edu.ifpi.ifala.services;
-
-public class DenunciaService {
-
-}
