+++ apps/ifala-backend/src/main/resources/application-docker.properties
@@ -1,0 +1,12 @@
+
+spring.datasource.url=jdbc:postgresql://postgres:5432/ifala
+spring.datasource.username=postgres
+spring.datasource.password=postgres
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+
+spring.jpa.show-sql=false
+spring.jpa.properties.hibernate.format_sql=true
+logging.level.org.hibernate.SQL=debug
+logging.level.org.hibernate.orm.jdbc.bind=trace
+
+spring.jpa.open-in-view=false
