+++ apps/ifala-backend/src/main/resources/application-prod.properties
@@ -1,0 +1,8 @@
+spring.application.name=ifala
+
+spring.datasource.url=jdbc:postgresql://localhost:5432/ifala
+spring.datasource.username=postgres
+spring.datasource.password=postgres
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+
+spring.jpa.open-in-view=false
